+++ src/lib/config.js
@@ -5,42 +5,38 @@ import _ from 'lodash';
 //
 
 // Possible values are STAGING, DEVELOPMENT, PRODUCTION
-// If you have connection error set as STAGING
 const environment = 'STAGING';
 
-const apis = {
-	hostname:
-		environment === 'PRODUCTION'
-			? 'http://localhost:3000/api/v1'
-			: environment === 'STAGING'
-			? 'https://mws-staging.ovr.ai/api/v1'
-			: 'http://localhost:3000/api/v1', // DEVELOPMENT
-	socket:
-		environment === 'PRODUCTION'
-			? 'ws://localhost:3000/cable'
-			: environment === 'STAGING'
-			? 'wss://mws-staging.ovr.ai/cable'
-			: 'ws://localhost:3000/cable', // DEVELOPMENT
-	creditCardApi:
-		environment === 'PRODUCTION'
-			? 'https://web3-helper.ovr.ai/buy'
-			: environment === 'STAGING'
-			? 'https://web3-helper-staging.ovr.ai/buy'
-			: 'https://web3-helper-staging.ovr.ai/buy', // DEVELOPMENT
-	indacoinBuyUrl:
-		environment === 'PRODUCTION'
-			? 'https://web3-helper.ovr.ai/indacoin'
-			: environment === 'STAGING'
-			? 'https://web3-helper-staging.ovr.ai/indacoin'
-			: 'https://web3-helper-staging.ovr.ai/indacoin', // DEVELOPMENT
-	etherscan:
-		environment === 'PRODUCTION'
-			? 'https://etherscan.io/'
-			: environment === 'STAGING'
-			? 'https://ropsten.etherscan.io/'
-			: 'https://ropsten.etherscan.io/', // DEVELOPMENT
-	cookieDomain: environment === 'PRODUCTION' ? 'ovr.ai' : environment === 'STAGING' ? 'ovr.ai' : 'localhost', // DEVELOPMENT
-};
+let apis;
+
+if (environment === 'PRODUCTION') {
+	apis = {
+		hostname: 'http://localhost:3000/api/v1',
+		socket: 'wss://localhost:3000/cable',
+		indacoinHelperApi: 'https://web3-helper.ovr.ai/indacoin',
+		etherscan: 'https://etherscan.io/',
+		cookieDomain: 'ovr.ai',
+		sumsubApi: 'https://api.sumsub.com',
+	};
+} else if (environment === 'STAGING') {
+	apis = {
+		hostname: 'https://mws-staging.ovr.ai/api/v1',
+		socket: 'wss://mws-staging.ovr.ai/cable',
+		indacoinHelperApi: 'https://web3-helper-staging.ovr.ai/indacoin',
+		etherscan: 'https://ropsten.etherscan.io/',
+		cookieDomain: 'ovr.ai',
+		sumsubApi: 'https://test-api.sumsub.com',
+	};
+} else {
+	apis = {
+		hostname: 'http://localhost:3000/api/v1',
+		socket: 'ws://localhost:3000/cable',
+		indacoinHelperApi: 'https://web3-helper-staging.ovr.ai/indacoin',
+		etherscan: 'https://ropsten.etherscan.io/',
+		cookieDomain: 'localhost',
+		sumsubApi: 'https://test-api.sumsub.com',
+	};
+}
 
 const map = {
 	lat: 46.0922495,
@@ -52,6 +48,7 @@ const map = {
 
 let config = {
 	web3network: environment === 'PRODUCTION' ? '1' : '3',
+	environment: environment,
 	apis: apis,
 	map: map,
 };