+++ src/views/Profile/Profile.js
@@ -2,16 +2,20 @@ import React, { useContext, useEffect, useState } from 'react';
 import * as moment from 'moment';
 
 import { UserContext } from '../../context/UserContext';
-import HexImage from '../../components/HexImage/HexImage';
+// import HexImage from '../../components/HexImage/HexImage';
 import HexButton from '../../components/HexButton/HexButton';
-import CheckBox from '../../components/CheckBox/CheckBox';
-import EmailConfirmation from '../../components/EmailConfirmation/EmailConfirmation';
-import IdensicComp from '../../components/IdensicComp/IdensicComp';
+import config from '../../lib/config';
+// import CheckBox from '../../components/CheckBox/CheckBox';
+// import EmailConfirmation from '../../components/EmailConfirmation/EmailConfirmation';
+// import IdensicComp from '../../components/IdensicComp/IdensicComp';
 import { getSumsubData } from '../../lib/api';
 import Blockies from 'react-blockies';
 
 import ValueCounter from '../../components/ValueCounter/ValueCounter';
 import { Web3Context } from '../../context/Web3Context';
+
+import snsWebSdk from '@sumsub/websdk';
+
 // import { networkError } from '../../lib/notifications';
 
 const ProfileContentLoginRequired = () => (
@@ -77,22 +81,49 @@ const renderBadge = (status) => {
 	return badge;
 };
 
+const launchWebSdk = (apiUrl, flowName, accessToken, applicantEmail, applicantPhone) => {
+	let snsWebSdkInstance = snsWebSdk
+		.Builder(apiUrl, flowName)
+		.withAccessToken(accessToken, (newAccessTokenCallback) => {
+			// Access token expired
+			// get a new one and pass it to the callback to re-initiate the WebSDK
+			let newAccessToken = '...'; // get a new token from your backend
+			newAccessTokenCallback(newAccessToken);
+		})
+		.withConf({
+			lang: 'en',
+			email: applicantEmail,
+			phone: applicantPhone,
+			onMessage: (type, payload) => {
+				console.log('WebSDK onMessage', type, payload);
+			},
+			customCss: 'url',
+			onError: (error) => {
+				console.error('WebSDK onError', error);
+			},
+		})
+		.build();
+	snsWebSdkInstance.launch('#sumsub-websdk-container');
+};
+
 const ProfileContent = () => {
 	const { state: userState } = useContext(UserContext);
 	const web3Context = useContext(Web3Context);
 	const [sumsubShowPanel, setSumsubShowPanel] = useState(false);
-	const [sumsubExternalUserId, setSumsubExternalUserId] = useState(0);
-	const [sumsubAccessToken, setSumsubAccessToken] = useState(0);
+
 	const { user } = userState;
 	// console.log('ProfileContent->These are user and web3Context from UserContext', web3Context, user);
-	console.log(web3Context);
 	useEffect(() => {
-		if (sumsubShowPanel == true && sumsubExternalUserId == 0 && user.uuid != undefined) {
+		if (sumsubShowPanel == true && user.uuid != undefined) {
+			console.log('eccolo 2');
 			getSumsubData()
 				.then((response) => {
+					console.log(response);
 					if (response.data.result === true) {
-						setSumsubExternalUserId(user.uuid);
-						setSumsubAccessToken(response.data.content.token);
+						console.log(user.email);
+						launchWebSdk(config.apis.sumsubApi, 'basic-kyc', response.data.content.token, user.email, null);
+						// 	externalUserId={sumsubExternalUserId}
+						// 	accessToken={sumsubAccessToken}
 					}
 				})
 				.catch((error) => {
@@ -145,15 +176,12 @@ const ProfileContent = () => {
 							</div>
 						</div>
 					</div>
-					{sumsubExternalUserId != 0 && (
-						<IdensicComp
-							baseUrl="https://test-api.sumsub.com"
-							externalUserId={sumsubExternalUserId}
-							accessToken={sumsubAccessToken}
-						/>
-					)}
+					<div className="p-tiny-message">
+						{user.kycClientComment != null && user.kycReviewAnswer != 1 && user.kycClientComment}
+					</div>
+					<div id="sumsub-websdk-container"></div>
 				</div>
-				<div key="redeem-land" className="p-section --m-t">
+				{/* <div key="redeem-land" className="p-section --m-t">
 					<h3 className="p-section-title">Redeem lands</h3>
 					<div className="p-section-content">
 						<h4 className="p-content-title">Waiting to be redeemed</h4>
@@ -169,14 +197,7 @@ const ProfileContent = () => {
 							</div>
 						</div>
 					</div>
-					{sumsubExternalUserId != 0 && (
-						<IdensicComp
-							baseUrl="https://test-api.sumsub.com"
-							externalUserId={sumsubExternalUserId}
-							accessToken={sumsubAccessToken}
-						/>
-					)}
-				</div>
+				</div> */}
 				{/*
 					<div key="notifications" className="p-section">
 						<h3 className="p-section-title">NOTIFICATIONS</h3>